requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.0.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
numpy==1.24.2
openai==0.26.5
pandas==1.5.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
tqdm==4.64.1
urllib3==1.26.14
yarl==1.8.2
